Started work on making support libraries
FOllowing D. Vallado's idea started to work on moving the contents of
stMat, astTime etc. to dynamically linked libraries (DLLs). It appears
impossible to link one DLL from another, so some of them need to be
compiled as static libraries (astMath so far) to be linked from the DLLs
(astTime so far). Current configuration compiles both in debug and
release configuration but wok still needs to be done in order to convert
other functions.